Remove using status field to sync cloudmap service for virtual-node
(TIL) Using status field to decide whether to take action or not is an anti-pattern. This is causing bug where, if user deletes a Service on Cloud Map side using AWS tooling, then controller is not creating the service again. This results in pods which do not get registered with Cloud Map due to stale service-id.
WS Cloud Map APIs do not fail synchronously i.e. RegisterInstance will not fail if service is missing, instead it returns an Operation response that will eventuall fail.

This change ignores the service-id in status field (informational purposes only) and calls CreateService. Note that it does not call GetService because CreateService is cheaper.