Modifies integration tests to be runnable outside of AWS
This patch reworks the test and CI system for integration tests to make the
execution of the tests more platform agnostic. A system to manage and inject
credentials is added that makes the tests runnable outside of an AWS context,
and substantial adaptations are made to the test setup and runtime to
make them run on a pure github action or linux environment.