Deploy a multiarch image
Previously, the "main" image was always an amd64 image, and there was
a secondary arm64 image that could be used with an explicit tag. This
makes the amd64 image also have an explicit tag, and replaces the normal
tagged image with a manifest file that will correctly route to the appropriate
architecture.