Fix VirtualRouter route deletion sequence (#768)
* Remove routes before the matching listener

This addresses a few rough edges with updating a VirtualRouter. After
this change, it is possible to update the listener protocol on the CRD
in one step. It is also possible to remove the listener at the same time
as the routes which use that listener.

* Fix a nil panic error

When running this against a test cluster, discovered that the logic must
filter to only names with SDK routes before adding them to a deletion
list.

* Add unit test coverage

Adds unit test coverage for the new functions. A possible nil pointer
dereference is fixed, which could occur if port matching is not
utilized.

* Remove unneeded API calls

The API call to describe the route is not needed, as the route ref
contains all the information needed to delete routes.